% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RClassification.R
\name{s.holdout}
\alias{s.holdout}
\title{Leave-one-out (Lachenbruch "Holdout") Procedure for equal and unequal variance}
\usage{
s.holdout(x, y, i, var, g)
}
\arguments{
\item{x}{is a 1st Group}

\item{y}{is a 2nd group}

\item{i}{is such a variable which you went to test whether it from 1st or 2nd Group}

\item{var}{is use to select a method whether it has for equal or unequal variance (Var="Equal" for equal and Var="NotEqual" for unequal variance)}

\item{g}{is use to define a group i.e the test value is belong from 1 group or from 2nd group (g=1 for 1st Group and g=2 for 2nd group)}
}
\description{
Leave-one-out (Lachenbruch "Holdout") Procedure for equal and unequal variance
}
\author{
Syed Hammad Raza
}
